requirements.txt
django==4.1.3
django-autoslug==1.9.8
djangorestframework==3.14.0
requests==2.28.1